changed CarController, Request, Model, Migration, api route, added hasMany relationship in Shop model
