+++ app/Http/Requests/CarCreateRequest.php
@@ -24,12 +24,13 @@ class CarCreateRequest extends FormRequest
     public function rules()
     {
         return [
-            'name' => 'required|min:2|max:255|unique:cars,name',
+            'name' => 'required|min:2|max:255',
             'price' => 'required',
             'condition' => 'required',
             'year' => 'required',
             'color' => 'required',
             'speed' => 'required',
+            'shop_id' => 'exists:shops,id',
             'category_id' => 'exists:categories,id',
             'brand_id' => 'exists:brands,id'
         ];
